Splitting translation units and adding namespace
